rename: 지역 좌표 파일들 이름 변경
