+++ src/assets/meteorologicalRegionCoordinates.ts
